Added "Empty" Element.
You can use the Empty element in the element array as "new Element(string text),".

When it is drawn it is executed as string and does not create an element.

You can get the same effect by creating an element with the optional parameter set to true, like this:

new Element(null, "Text", null, true),